axyo-engagement-app/data-refinery-app/requirements.txt
# requirements.txt
flask
pandas
python-dotenv
flask-wtf
numpy
aiohttp
requests
PyYAML
faker